+++ Cargo.toml
@@ -2,3 +2,7 @@
 name = "b-compiler"
 version = "0.0.1"
 edition = "2021"
+
+[dependencies]
+anyhow = "1.0.68"
+config-manager = "0.1.0"
