+++ src/lib.rs
@@ -1,0 +1,5 @@
+mod config;
+mod intermediate_code_generator;
+mod machine_code_generator;
+mod parser;
+mod tokenizer;
